ABZU: Fix common problem with new and improved pattern
These are more instances of the common pattern we are seeing everywhere
in this game, but using a new and improved version of the fix. Rather
than inversing the SVPositionToTranslatedWorld matrix and needing to
find the render size from another constant buffer, we use the
TranslatedWorldToClip and ClipToTranslatedWorld matrices instead.