ABZU: Fix accidental hard coding of resolution (UNTESTED)
This was one of the common patterns, but as it does not operate on
SVPositionToTranslatedWorld in the FViewUniformShaderParameters constant
buffer I can't be sure if the new pattern will work. Taking a gamble
as to which constant buffer has the resolution based on a second
adjustment I had commented out - need to find the shader and verify.