ABZU: Add python script to autofix common pattern
This isn't integrated into hlsltool yet, and given this game has 10,000
shaders that could match I'm reluctant to run it globally and risk
introducing decompiler bugs, but these kind of simple patterns is where
I want to take hlsltool going forward (i.e. move the boilerplate to
hlsltool so the individual game scripts can focus on the patterns),
reserving asmtool for more advanced fixes or where a high degree of
dependability is required.