asmtool: Move lval conversion to register to a lower level
We reused code from hlsltool to parse a basic Register, but hlsltool has
to contend with the lval being almost anything, whereas asmtool knows
that it is a register of some sort. Perform the string to register
conversion in the asmtool Register class so we can remove it from the
individual patterns.

We no longer subclass hlsltool's AssignmentInstruction class because the
only overlap left is the reads() routine - writes() now uses a different
implementation.