asmtool: Stop ResourceLoadInstruction matching sample_c_lz_aoffimmi_indexable
The ResourceLoadInstruction pattern was matching lines like:

    sample_c_lz_aoffimmi_indexable(0,0,0)(texture2d)(float,float,float,float) r0.w, r2.yzyy, t3.xxxx, s2, r2.w

The third set of brackets in this instruction was not handled correctly
by this class resulting in lval="(float,float,float", breaking the new
assumption that lval is a register.

However, these are not actually resource load instructions, and should
just be left as generic AssignmentInstructions unless we have a need to
parse them later, so modify the pattern to reject lines with a third set
of brackets.