asmtool / hlsltool: Add a stop_when_clobbered option to scan_shader
This is intended to be used when searching for a register access if we
want to stop if the register is overwritten. The obvious case is where a
register is overwritten when we are searching for a read, but also
handle the case if it is overwritten while searching for a write of a
specific instruction.

Keep it disabled by default for now just in case we have any
dependencies on the old behaviour, but we probably do want to switch
this to the default as it simplifies searching.