WATCH_DOGS2: Fix depth of light halo on distant tower
This one is interesting - the output position doesn't have depth, but we
recognise a pattern in the shader is similar to the unprojection pattern
which uses the inverse projection matris to find the linear depth. Using
that value as the depth in the stereo correction formula moves the light
to the correct depth.