WATCH_DOGS2: Unbind the depth buffer when drawing the HUD
It seems this game has a depth buffer bound while drawing the HUD, which
confuses the driver heuristics causing it to sometimes adjust the HUD,
and other times not. I saw this change while playing, so instead of
trying to guess it's decision, unbind the depth buffer when drawing the
HUD which will influence it to never adjust the HUD.

Need to keep an eye out if this might break anything in case there was a
reason the game had the depth buffer bound (like using a stencil buffer
to mask off part of the HUD).

I also tried normalising the output coordinate, but performing the
perspective divide in the shader causes massive artefacts for some
reason (floating point error? w is very large, as in 1e38!).