WATCH_DOGS2: Fix volumetric fog around moon
This volumetric fog shader is a bit special - it is responsible for the
glow around the moon (and possibly haze on the horizon, but I can't see
that yet outside of the debug shader) and combining some of the other
volume buffers.

The autofix does not adjust the glow around the moon, so apply a simple
change of adjusting the screen position that has the desired effect.
This seems ok right now, but may need more work if it causes issues with
horizon haze or anything else.