WATCH_DOGS2: Synchronise falling leaves between both eyes
The falling leaves do not line up between both eyes, and will even
settle on the ground in completely different locations. This is due to
the physics being calculated in a compute shader using the depth buffer,
but since the depth buffer is slightly different in each eye the physics
can desync.

I tried a number of ways to force either the leaf structured buffer or
depth buffer to be mono, but nothing seemed to work. Instead, just
unbind the depth buffer from the physics shader so that the physics will
match up between each eye. It does mean that the leaves won't settle on
the ground, but that is a minor issue that will probably go unnoticed.