asmtool: Add pattern for WATCH_DOGS2 screen space reflections
Seems there are quite a few variants of this shader, so script the fix
to catch them all. Compared to the hand fixed version, this one does not
apply the unprojection fix to all cases where it could apply - only the
first seems to have a noticeable impact and the others would need that
autofix to be adapted to work with slightly different instructions.