asmtool: Fix regression in --auto-fix-vertex-halo
When I changed the class hierarchy the other day I forgot that the auto
halo fix was implemented in hlsltool and required that the asmtool
AssignmentInstruction inherited from the hlsltool AssignmentInstruction
so that it's isinstance() would pass.

Turn it back into a subclass, just don't call the super class __init__
for reasons described previously. Also make
expression_as_single_register accept a register and return it
immediately so that it will work with the new changes in asmtool.