WATCH_DOGS2: Add custom shader to debug transparent depth buffer issues
This shader will overlay depth information in the world for any objects
on the transparent depth buffer, making it easy to spot broken shaders
that are otherwise difficult to see.

The depth buffer is obtained via the confetti physics shaders, which is
where I noticed this problem, but have confirmed that it affects
reflections of these transparent objects.