WATCH_DOGS2: Fix flares in cinema + fix flare visbility
These use the same shader as the sun flare, but these have the correct
depth available, so switch to using that (I'll need to double check this
didn't break the sun flare later).

The light visibility is calculated in a separate shader tracked down by
following the buffer through frame analysis. Adjust the position in that
shader as well to fix the visibility.