shadertool/asmtool/hlsltool: Print vanity comment to stderr
My patterns are getting more terse as time goes on, and I often don't
bother printing a message to stderr that the tool is applying a fix, so
have the vanity_comment function print this message instead.

I should clean up the duplicate messages, but that is a lower priority.