asmtool: Fix volumetric fog fix failing on early CameraPosition.z access
Some of the volumetric fog shaders calculate the height above sea level
using CameraPosition.z, which would cause the volumetric fog fix to
match, but fail to apply when only one component of CameraPosition was
accessed.

Change it to iterate over all accesses of CameraPosition, only applying
when three components are accessed.