asmtool: Add support for desmearing shadow volume shaders
The shadow volume shaders follow a slightly different pattern to the
other volumetric fog shaders, and need the original position preserved
to access the volume from the previous frame - otherwise it is smeared.
Add support to search for the texture used to access the previous frame
and switch it back to the original register.