WATCH_DOGS2: Add note about which shaders fog affects
The fog constant buffer is not by itself enough of an indication as to
which shaders we need to apply the fog pattern to, and the answer is not
obvious. To try to narrow down the possibilities, add some notes to
track which shaders have the constant buffer, whether the pattern
applied to them or not, and what they are for (if known).