asmtool: Add alternate pattern to fix WATCH_DOGS2 volumetric fog
Rather than adjusting a position part way through the shader, this
adjusts the inputs to the shader to account for the off-axis rendering
used in stereoscopic rendering. It adjusts the camera Z axis to match
separation (the off-axis part is independent of convergence and
essentially sets where infinity lies - same thing happens in a real
camera when using a tilt/shift lens to move the lens off-center).

It needs to be used in conjunction with the view direction
reconstruction fix, and relies on the fact that the corrected
view/camera direction now matches the stereoscopic projection, so the
volumetric shaders will use the stereo corrected direction when
projecting away from the camera to sample the fog volume.

The results are slightly different to using the regular fix, but it is
hard to say if they are better or worse.