WATCH_DOGS2: Adjust part of camera zoom HUD
The circle on the camera zoom HUD was already adjusted (same texture as
the camera focus I guess). This adjusts the bar as well. The + and -
text is not adjusted - did a quick check and it uses a common font
texture, so it would need more advanced filtering, but it's not worth
it.