WATCH_DOGS2: Force transparent depth buffer to stereo, remove 2D glass shaders
In non-SLI the transparent depth buffer was mono, so the stereo
correction was only shifting the reflection of glass objects to the
right. Forcing the buffer to stereo caused the driver to apply it's
stereo correction, resulting in a double correction, so remove our
manual correction.

Need to check if this breaks SLI - if it does we will need to neutralise
the stereo correction.