WATCH_DOGS2: Fix the single GPU left eye HUD flicker
Some of this is speculation:

The final render target the game uses in the immediate context is for the
HUD, but it is *NOT* the back buffer, nor a texture (at least not that
3DMigoto ever saw get created), and the bizzare thing is - we never see any
kind of copy operation to draw it to the back buffer!

This suggests that that copy operation is somehow happening out of band
(not entirely sure how, maybe using a different device? certainly in another
thread), and that only waits until the game has submitted it's final draw
call before starting the copy - but it does not realise that the 3D Vision
driver is duplicating the draw calls and is still submitting the final few
calls for the left eye. This could be a driver bug if a draw call is allowed
to complete before the driver has duplicated it, but they probably have their
reasons (performance) that doesn't excuse the game copying this render target
out of band.

In order to fix this, we redirect the HUD to our own custom render target,
which we will then copy to the back buffer from the [Present] call
(essentially, are are doing what the game was supposed to do).