WATCH_DOGS2: Fix lighting shaders used in NVIDIA Ansel
I just discovered that Ansel allows the game to be paused with a broken
effect displayed, and allows the camera to be adjusted to get a better
view - useful to track down effects that are only displayed for a short
time, or while under pressure. Only problem is that some of the effects
switch to different shaders while in Ansel - most notably shadows.

This fixes a couple of light/shadow/specular shaders in Ansel so they
aren't as distracting and we can focus on the effects of interest. These
shaders are not among those dumped when loading the game, nor were they
among those extracted with generic_shader_extractor.