WATCH_DOGS: Fix clipping and uneven light on outside of police search light shaft
The clipping was coming from the driver steroising one use of one of the
light shaft's vertex shaders, but not the other. Adding our own
correction double corrected the one the driver had corrected, so we
neutralise the driver's correction using the cb12 technique (normalising
the output coordinate removed the effect).

The uneven light on the exterior of the light shaft was tracked down to
one of the pixel shaders that contribute to the effect taking it's view
space coordinate and projecting it to a different depth - adding a
stereo correction before the projection, then removing it afterwards
corrects this.

There is still some minor clipping occurring when the camera passes
through the boundary of the light shaft - just note that for now.