WATCH_DOGS2: Fix SSR when temporal filtering is enabled on a single GPU
This specific combination leads to broken SSR reflections, due to a mono
depth buffer sneaking in. The buffer is mono because it is drawn to with
no render target or pixel shader assigned. Assign a dummy render target
when it is being drawn so the driver will steroise it, and fix the
reflections.

n.b. StereoTextureEnable = 0x400 is supposed to fix these types of
issues, but doesn't ever seem to work.