WATCH_DOGS2: Return throw arc to correct depth
The throwing arc uses the same shader pair as the waypoints on the
minimap, but we don't want to return it to screen depth. Filter based on
the texture being used (other filter types have problems - render target
size filtering doesn't work so well because pixel density and temporal
filtering can create a difference in resolution size and render target
size, and the depth buffer is active in all cases).