debug_volume: Support scaling the volume slices
This debug shader was originally written for the volumetric fog in
FCPrimal, which had a lot of small slices and I hardcoded a few values
to get it working. It can be nice to view other volumes that may not
have as many slices and scale up those it does have.