WATCH_DOGS2: Fix brickwork decals - new stereo correction technique
These don't follow the same pattern as the other decals, and I had
barely noticed there was anything wrong with them, but they do hover
slightly above the surface and shadows don't land on them.

This time round I'm trying a new technique - I modify the camera
position in the vertex shader, which in this game causes it to apply the
camera offset part of the stereo correction, so I remove the driver's
stereo correction and apply only the off-center part of the projection.
With a bit of algebra that simplifies to just adding
separation*convergence to the output coordinate:

 x' = x - s * (w - c)
x'' = x' + (s * w)
x'' = x - s * (w - c) + (s * w)
x'' = x - (s * w) + (s * c) + (s * w)
x'' = x + (s * c)

With this technique it lines up perfectly with the pavement without
having to understand the details of exactly what the shader is doing.

Variations of this technique may work in other games if either the
camera position or view matrix is used in the vertex shader to calculate
the position.