WATCH_DOGS2: Fix clipping on thick San Francisco fog
The clipping was being caused by a mono depth ranges texture, and adding
the stereo correction to the vertex shader that draws that fixes the
clipping, but breaks the fog lighting way worse. Filter based on the
pixel shader to only adjust it when it is contributing to the depth, not
the lighting.