asmtool: Add 'FVCameraZAxis' to Z axis correction & allow negated correction
Adjusting the camera Z axis might just be the ticket to the remaining
fog issues, as it seems to be the only input that can adjust the sun
glow in the two ray marched fog shaders, and in fact combined with the
camera position fix it replaces the volumetric fog adjustment in those
shaders - suggesting that this might be the "correct" way to fix these
shaders... but it's a risky proposition to change the other shaders to
use the same at this stage.

It needs to be negated compared to my previous experiments, so add that
as an option.