WATCH_DOGS2: Match bullet texture on the GPU
Since we can't match the bullet texture from 3DMigoto without turning on
texture hash tracking (which incurs a significant performance penalty),
perform the matching for this texture on the GPU instead. Pass a copy of
the bullet texture extracted with frame analysis to the shader and have
it check if it matches the texture in ps-t0.

Note that in this case the texture is a typeless resource (with the type
set in the view), and the DDS file dumped with frame analysis inherits
the typless. When 3DMigoto loads the DDS it has to guess the type, and
guesses incorrectly in this case - it picks the _UNORM variant of this
type, but the game actually used _UNORM_SRGB. That leads to differences
in the colours and the match failing. For now, I just hexedited the DDS
file to set the correct type, but really this should be fixed in
3DMigoto to use the type from the view when dumping with frame analysis
(which implies ditching DirectTK and all it's limitations).