WATCH_DOGS2: Add convergence override to weapon scope
The scope is 3D and a little too close to the camera - usable, but not
ideal. Add a convergence override (similar to the 3D HUD), which seems
to work pretty well. This also adds the code to flatten it and move it
to crosshair depth, but I find I prefer keeping it 3D.

Adjusting this to crosshair depth is slightly complicated by the fact
that it is drawn earlier in the frame than anything we use to obtain a
reference to the depth buffer. Change it so that the copy of the depth
buffer is not set to null on present and will keep a copy of the most
recent valid depth buffer.