Revert "WATCH_DOGS2: Add convergence override to weapon scope"
This adjustment messed up the CTOS box in Prime_Eight's bunker. Pixel
shader filtering doesn't help. A depth threshold can work ok if it is
about 0.7 or so, but it is possible the camera might get pretty close
(certainly I can in camera mode), so for now just remove this
adjustment. Maybe we could adjust it based on the scope's crosshair
shader being active... later.

This reverts commit bf2d3373d2ee9201b9a7ce226f2b35ebf0fb28e7.