WATCH_DOGS2: Re-enable scope adjustment and trigger from scope crosshair
This is an attempt to allow the scope convergence override without
interfering with unrelated objects in the world. When the crosshair is
drawn in the scope we set a flag that is checked when adjusting the
scope itself so that the adjustment will only be active while the
crosshair is drawn.

Due to the draw order we can't clear the flag on the present call - the
crosshair is drawn late in the frame and the scope is drawn early.
Instead, we clear the flag after the scope shader has been called, so
only one object (hopefully the scope) will ever be adjusted each frame.
There will be one frame of lag, but it shouldn't be noticeable.