WATCH_DOGS2: Remove leaf kill switch, add very low convergence preset, update README
There haven't been any further reports of leaves desyncing, so remove
the kill switch and replace it with a very low convergence preset for
use in ATM cameras, satellites, etc. where the regular convergence
presets are too high. I didn't want this on the main toggle key, because
it is far too low for regular gameplay and I like switching on the fly
when entering vehicles.