WATCH_DOGS2: Note that we need a resource pool in 3DMigoto
As of the 1.09 update, the log file is showing this over and over:

RecreateCompatibleTexture: Recreating cached resource

This is coming from the dummy render target assignment used to stereoise
reflections. In 1.09 they optimised this by using a downscaled depth
buffer, and this shader can now be called with render targets of varying
sizes. This causes 3DMigoto to constantly destroy the old cached render
target and recreate it to match the depth buffer size.

This works alright, but is wasteful. Ideally 3DMigoto would keep a pool
of cached resources that it could choose an appropriate resource from.