WATCH_DOGS2: Disable white light glitch fix
This fix still causes problems - it creates some massive artefacts
visible through the walls of an air vent in the final mission. Both
glitches are pretty situational, but in this case I think the airvent
glitch is worse, so disabling the fix for now. Later we might add a
toggle for the fix, but that will need conditional logic in the
d3dx.ini to work.