Add script to convert images taken with the Loreo 9005 lens to .jps
Script to convert photos taken with the Loreo 9005 3D DSLR lens in a cap
into .jps format. This script is not intended to be used alone - the
.jps files should be cropped with my Stereo Photo Cropping Tool
afterwards to adjust the parallax and/or crop off the overlap. By
default this tool will copy 5% overlap from the center of the image to
both eyes - the amount can be adjusted, but keep in mind that since the
divider between eyes is not in a fixed position on this lens (and there
is no way to automatically determine what it was set to), you should be
generous.