Add extract_wd2_shaders script
This script is a first pass at extracting shaders for WATCH_DOGS2
directly from the patch.dat file. Shaders in this file are compressed
using LZ4, so the script searches for the DXBC marker, which due to the
way that LZ4 works is likely to still be in plaintext, then it
backtracks from there searching for something that looks like a valid
LZ4 stream.

The DXBC shader starts some way into the LZ4 stream, so it searches the
decompressed data for the header and tries to decode it. It will use the
filesize from this header to determine when to stop decompressing since
the LZ4 stream lacks an end of stream marker.

Compared to the Unity implementation of LZ4, this seems to have a few
more edge cases (though it would be worth going back and see if they
apply to the Unity version as well) - namely that the match offset field
can be extended if the three highest bits are all set, and that the end
of the stream can be repeated if the match length is longer than the
match offset.

Most of the shaders are extracted successfully with this approach,
though several throw errors while trying to determine the shader model
that will need checking, that suggest the chunks field in the header is
too high.