extract_wd2_shaders: Verify hash
Since we are decompressing the shaders without knowing the correct
starting offset we can sometimes decompress something that looks like a
valid shader, but is not. Check the embedded hash to discard any that
are not valid.

What tends to happen is the correct shader is immediately extracted
after discarding the incorrect shader, as the script attempts to start
the decompression from an earlier byte.