dx11shaderanalyse: Pre-calculate the sines of integers for the shader hash
Doing the shader hash calculation in Python is slow - it is not a
calculation well suited to Python, and would be better of being
implemented in a C extension. I generally try to avoid using C
extensions since they complicate installation on Windows though, so for
now just grab some low hanging fruit and move some of the static
calculations out of the function to they are only run once.

The speedup is pretty marginal, which is to be expected. I could
potentially unroll the main loop, but I also don't want to make the
procedure harder to read, so leave that alone.