mkrelease: Add support for releasing as 7z files
The next release of the WATCH_DOGS2 fix would be a 150MB zip file (rar
is 23MB if compressing as a solid archive, or 143MB as a regular
archive, but rar is not free, so I'd rather avoid it), but as a solid 7z
file it is only 7MB due to the superior compression algorithm and it's
ability to use the redundancy between different files in the archive, so
add that as an option to the mkrelease script.

I also experimented a little with tar.xz, but 7-zip on Windows shows the
contained tarball, which is not ideal. 7zip seems more straight forward
for Windows users.

Refactor the argument parsing to allow --7z to be specified anywhere
on the command line.