hlsltool/asmtool: Fix edge case in auto halo fix
The auto halo fix would miss some cases where the temporary variable was
written to on the same line it was read from - scan that line for reads
as well to catch these cases. This fixes some halos in heat distortion
effects in Dreamfall Chapters.

By itself this change became overzealous and started adding halo fixes
to various shaders in Dreamfall Chapters where it wasn't needed if the Z
component was read in that line. Restrict it to scanning for reads of
the X component to eliminate these.