SLI Optimised version of the 3DVision2TB shader
This is a variant of the 3DVision2SBS/TB shader that should improve
performance in SLI systems at high resolutions (above 1920x1080). It
works by performing two passes - the first pass downscales the back
buffer to half height on each GPU, and the second pass performs the
reverse stereo blit on the downscaled buffer then modifies the output to
top-and-bottom. Since the reverse stereo blit is the expensive part of
this operation, this should reduce it's bandwidth requirements over the
SLI bridge by half allowing for higher resolutions before this becomes
an issue again.

Side-by-side and line interlaced are not yet supported - SBS would need
a different sized buffer, which would be easier if 3DMigoto can be made
to automatically select between the two (or more to the point - only
perform the reverse stereo blit on the appropriate buffer for the
current output mode). Line interlaced could use the same half height
buffer and would be relatively straight forward to add.