asmtool: Add assembly version of --fix-unity-lighting-ps
The HLSL version of the lighting pixel shaders in Dreamfall Chapters
have an enormous impact on the framerate, dropping from 50fps to 38fps,
and most of this overhead is purely a result of decompiling and
recompiling the shaders.

Implement the pattern to fix Unity lighting pixel shaders in assembly to
recover this lost performance. This isn't exactly the same as the HLSL
version, since it skips using the FOV passed from the vertex shader (for
now - later when 3DMigoto supports assembling the ISGN section we can
add that), but otherwise follows the same procedure. Some of the helper
functions were refactored to be callable from asmtool.