Unity 5 / DX11: Precalculate clip to world matrix to skip local
For a few of the patterns we convert clip space coordinates into some
local coordinate scheme, then from there convert to world space. Rather
than doing this in every shader with these patterns, pre-multiply the
two matrices together in the inverse_unity_mvp shader so that it can be
done just once and we can remove a lot of the intermediate calculations.