Unity 5 / DX11: Improve performance by limiting full matrix copies
We perform a full copy of the matrix buffer for every shader that uses
it, and although individually this is negligible, it can add up. In
Dreamfall Chapters, this totals about a 5fps hit, so would be nice to
eliminate.

This patch reduces it to just two full copies per frame (or one if you
have removed any remaining dependencies on Resource_Inverse_MVP) that
are performed after calculating the buffer in the compute shader. This
resulting constant buffer is then simply referenced in each destination
sahder.

In combination with switching the lighting fix to assembly, this easily
gets us within 1 or 2 fps of the performance observed while holding the
show_original key.