asmtool: Add --remap-cb option
This option remaps accesses of a given constant buffer to use the same
location in a new structured buffer. This is intended to facilitate
injecting stereo corrections in buffers instead of modifying the
instructions, and since constant buffers can only be mono (with the
exception of the driver's stereo constant buffer) this allows us to use
a stereo structured buffer instead.

Since structured buffer accesses are slower than constant buffer
accesses it might be worth extending this to indicate exactly which
entries in the buffer need to be remapped so that the faster constant
buffer is still used for mono data, but for an initial proof of concept
this should work fine.