Unity5 / find_transparent_shaders: List shaders with duplicate hashes
We might sometimes need to remove shaders that this script identifies,
but since shader hashes can be shared between shaders we could
potentially miss some. Change the script so it doesn't filter out
duplicate hashes, but lists them for each shader they were found in with
notes as to which other shaders they were also found in. Only the first
ShaderOverride for each hash will be active - the rest will be commented
out.