asmtool / Unity lighting pattern: Use FOV passed from vertex shaders if available
This brings the lighting fix in asmtool up to par with the version in
hlsltool, using the FOV passed from the vertex shader if it is available
and only using the passed in inverse view-projection matrix if not.

This makes the lighting fix more reliable since if the FOV has been
passed from the vertex shader we have more confidence that it is
correct, whereas the inverse view-projection matrix could be thrown out
if it was copied from somewhere bogus.

Note that while this does amend the input signature comment, 3DMigoto
does not yet parse that section (cmd_Decompiler does), so the
reliability of this is questionable.

This also restores the Unity 5.3 autofix to use it

This also optimises the lighting and reflection fix patterns slightly