asmtool: Have --remap-cb handle constant buffers larger than 2048 bytes
Structured buffers have a maximum stride of 2048 (well, it seems we can
still assemble larger strides and the shader I'm looking at *seems* ok,
but let's play it safe). If we need to access something above that use
the index as the higher bits.

Also sort the constant buffer when remapping it so our inserted code
and register choices will be consistent and easier to read.