CB Debug Shader: Add an way to examine other buffers, including stereo buffers
This debug shader is particularly interesting because since it decodes
the buffer on the GPU it can see the different values in each eye of a
stereo buffer.

However, there is no known way to have a constant buffer be stereoised
in 3D Vision (with the exception of the constant buffer the driver uses
internally), so copying a buffer into [ResourceDebugCB] will lose the
values from one eye. This adds a [ResourceDebugBuf] that can be used
instead, and if it is set the shader will automatically switch to
displaying values from it instead of the constant buffer.

The bind flags on [ResourceDebugBuf] are overridden to include UAVs,
since I determined that it is not sufficient to force the driver's
surface creation mode to get a stereo buffer, but this bind flag will do
the trick.