UE4: Bring my stereo injection matrix POC in and clean it up a little bit
This adds a custom compute shader from my old proof of concept, updated
to produce a complete copy of the constant buffer that will be
stereoised. Only SVPositionToTranslatedWorld is currently stereoised, as
that has already been tested in the proof of concept, but all going well
we should be able to stereoise every relevant matrix in the buffer.