Senua: Fix depth buffer ray-traced shadows with stereo constant buffer technique
This is the realisation of my stereo constant buffer proof of concept.
It uses a compute shader to modify the FViewUniformShaderParameters
constant buffer from the game and inject stereo corrections into a bunch
of matrices, then pass the resulting buffer back to the original shaders
in a shader resource slot. The shaders are modified with asmtool to
redirect accesses from the original constant buffer to the modified
version, which gets around the driver's mono constant buffer limitation.

The custom shader is limited to running once per frame, because frame
analysis shows that this constant buffer is only updated once per frame.
Note that the FViewUniformShaderParameters differs slightly from stock
UE4 as the developers have added an additional matrix and vector in the
buffer, which can be identified with the debug shader or frame analysis.

This specific effect is not perfect in 3D, but neither is it in 2D -
it's similar to screen space reflections in a way, but for shadows of
plants and rocks, presumably to save the performance cost associated
with drawing them on a shadow map.