asmtool: Use exit code to indicate if a shader was installed
This will allow the calling script to check if asmtool applied a fix or
not e.g. to decide whether to update the d3dx.ini or not.

This also sets --remap-cb to set the autofixed flag if successful,
though intentionally does not set the flag on
--disable-driver-stereo-cb, since that is most likely going to be used
with some other change, and if it is not it can just be run
unconditionally over all vertex/domain shaders.