Senua: Fix volumetric lighting / light shafts
Well, this effect looks *very* familiar - I am quite certain that this
is the exact same volumetric lighting used in Far Cry Primal and
WATCH_DOGS2. Only difference here is that since this is UE4 the headers
have been stripped so asmtool can't follow it, but no problem - let's
use my new constant buffer stereoisation technique instead :)

Since this isn't a standard UE4 effect it doesn't use the standard UE4
constant buffer, so we have to analyse the constant buffers it is using
and match the matrices and camera position up, then use custom shaders
to copy the stereoised versions produced by my UE4 stereoisation shader.

This fixes light shafts, fog volumes, etc. There might be more shaders
to find (and if it's like WATCH_DOGS2 they may *not* all be compute
shaders), but with no way to extract and match them up we will have to
find them as we go.

Note that it was *very* important that the transformation matrices from
the previous frame be adjusted as well.