Constant Buffer Debug Shader: Automatically disable display if no debug resource set
The CustomResourceBuf we could easily check from HLSL with
GetDimensions, but there was no way to check if constant buffers were
bound... until 3DMigoto 1.2.65 that is, which extended texture filtering
to work with other resource types and added a method to check for
unbound resources. This now disables the text if the debug resources
have not been set to anything.