Senua: Fix directional lighting with constant buffer stereoisation technique
This fixes the directional lighting with the constant buffer
stereoisation technique. These ones have a screen to light matrix which
we need to stereoise. We don't already have a suitable stereo injection
matrix that operates in screen space, but it is easy to create one from
the mono and stereo versions of the ScreenToWorld matrix.