UE4/remap_cb.cpp: Add regular expression shadow fix
Some of the UE4 shadow shaders use a per-draw ScreenToShadowMatrix.
While we can use an inverse screen-space stereo injection matrix
(e.g. derived from the stereo and mono versions of ScreenToWorld) to
inject a stereo correction into that matrix, it is wasteful since it is
only used for one draw call, and uncertain how reliably we can detect
the buffer it is in since AFAICT it is just the shader's $Globals
buffer.

Instead, let's fall back to more traditional techniques for these
shaders and perform a regular expression search and replace on their
assembly to inject a stereo correction. For now the pattern is pretty
precise, but that can be relaxed as we find more variants of these (we
might even be able to get away with only matching the matrix multiply).

This also gave me a chance to start playing around with the regex
implementation in C++11, which barring any MS stupidity (have hit this
last time I looked at regex in MSVC) we might be able to adapt to some
form of more general autofix engine in 3DMigoto. If there is stupidity
to be found I might end up pulling in something like PCRE...