Senua: Add SLI profile
This is based off the SLI profile from:
https://www.forum-3dcenter.org/vbulletin/showpost.php?p=11455083&postcount=2622

I did my own experimentation on top of that, and found a variant that
does not break light shafts, but still gives good fps.

There is a nasty flicker when starting the game (in either variant), but
it clears up by itself after about 10 seconds, and the fps boost seems
worthwhile.