Game List: Reduce number of posts requested at a time to 100
Blogger started returning 500 errors a few days ago for no apparent
reason. As a stop-gap measure, reduce the number of posts requested at a
time to 100, which reduces the error rate from around 50% to more like
10%.

Also disable sorting the results on the server in case that is causing
the server to perform more work - we sort them as they are inserted, so
that doesn't matter.