Game List: Raise maxResults back to 500
The maxResults doesn't seem to have much impact on whether or not the
request will succeed at the moment (it may have some), but more requests
definitely does increase the chance that at least one of them will fail,
so it seems we have better odds of loading the game list without retries
if we request more posts at a time rather than a larger number of
smaller requests.

Raise the maxResults back to the max the API allows (500) so that we can
load the game list in two requests if all goes well, and to minimise the
number of retries if things go belly up.