Game List: Remove old heartbeat error ping
This particular error ping will only trigger if the callback function is
called on failure, which will not happen on jQuery AJAX requests - it
did work before since collate_responses was called from a <script> style
REST callback, which happened on success or error, but now we've
switched the first query to fetch the number of posts this won't happen
anymore.

It's a bit of a shame, since the heartbeat ping we get from the jQuery
AJAX error doesn't include the server response making the logs even less
useful than they were before, but whatever - they weren't exactly useful
anyway.