extract_unity55_shaders: Small progress deciphering next data structure
This next data structure looks likely to be encoding the subshader
sections from the old .shader files. This is going to be tricky - some
patterns are emerging, but the overall structure including how to
determine entry lengths is still unclear. This decodes the tags
(uncertain if they are subshader or pass tags) if present, but little
else - I'll need more shaders to analyse this, and preferably shaders
where I still have an old .shader file to compare.