extract_unity55_shaders: Parse up to the end of the subshader metadata
Huge progress reverse engineering this format - now decodes and dumps
out all parameters up to the end of the subshader metadata, including
bindings and so on that we will want later. Not much left to decipher -
little more than the shaders and their names.

Still have to convert this into a format that the following scripts are
expecting, and obviously extract the shaders.