extract_unity55_shaders: Actually extract the shaders
The shader format in the compressed blob is near identical to the format
in Unity 5.3/5.4, just with one extra field, so call into the
extract_unity53_shaders script to actually perform this extraction.

This means we have pretty redundant bind info, but on the positive side
it means we don't necessarily have to pack up all the fields we've spent
ages decoding into a data structure to get things working, so for now
set extract_unity53_shaders to just skip using the classic headers. We
might still need one or two added that aren't in the 5.3 headers for
some things later, like maybe ZWrite.

This now extracts all shaders from Life Is Strange: Before the Storm