extract_unity53_shaders: Create data structures for classic Unity headers
This creates data structures that will be used to construct the classic
Unity headers based on the information parsed in the binary shader
format. For now these headers are bare, just containing the basic
structure, shader name and subshader/pass numbers.

Since most of the binding information we need is available in the Unity
5.3 style headers we already produce we won't need to fill these data
structures out entirely (though we could), but we will need some
information, such as the Tags to detect SHADOWCASTERs if nothing else.