Add unity_asset_bundle_extractor script
Life is Strange - Before the Storm has shaders inside asset bundles,
which use a different file format to the regular asset files. This
script is aimed at eventually being able to extract the shaders from
these files so we can autofix them as well.

So far it only parses the file header.