unity_asset_bundle_extractor: Extract contained files
This decompresses and extracts the files contained inside an asset
bundle. These appear to be further asset files, which should in theory
be able to be processed with unity_asset_extractor.

This can take a fair amount of time, so once this is working we might
consider changing to an on-demand block based extraction class.