unity_asset_extractor: Handle asset files with embedded type info
The asset files contained within asset bundles have a flag set that we
haven't seen before, which signifies that the type table contains an
additional data structure that appears to describe the fields in that
type. This may be useful to parse (and could have been useful to decode
the new shader format), especially if we come across a case where a game
is including asset bundles from multiple Unity versions, but for now we
don't really need it, so skip over the data structure so we can get to
the resources we actually want to extract.

Also add an optimisation to exit early if the type table doesn't list
any of the resource types we are interested in extracting.