unity_asset_bundle_extractor: Skip intermediate asset file extraction
Rather than extracting the asset files contained in the asset bundle
files, directly call into unity_asset_extractor with a stream for each
of the asset files to directly extract their contents.

For now this skips one step, but we should be able to optimise the
stream so that we only decompress blocks we actually need to make this
much faster.