unity_asset_bundle_extractor: Decompress blocks on demand
This introduces two layers to massively speed up extracting shaders from
asset bundles:

First, it introduces a BlockStream class, which exposes a file like
interface to access the decompressed data, but performs the
decompression on the fly only for blocks that are actually read. This
allows us to skip the vast majority of the time we would have to spend
decompressing the entire contained asset files.

Secondly, it introduces an OffsetIO class, which exposes a file like
interface that wraps another file like interface such as a BlockStream,
and applies an offset to the IO to allow us to map distinct files onto
the block stream. This is essentially akin to an OS filesystem layer.