LiSBtS: Add game-specific autofix.sh
This is copied from the Unity 5.5 autofix script, but with an additional
option to extract the asset bundles that this game uses.

The reason for not adding this to the generic script is because I don't
think that the ".bytes" extension that this game uses is standard, and
can't identify a fool-proof general way to identify asset bundles, and
processing these asset bundles can take a fair bit of time, and since
most games don't use them we may as well skip them in the general case.