unity_asset_bundle_extractor: Be more robust identifying UnityFS files
It's not so clear if the asset bundle files are going to be easily
identifiable from the filesystem alone in a way that works for all
games, so we might end up in a situation where we try the extraction
script on a file that isn't an asset bundle. Make the UnityFS signature
detection a little more robust to better cope with random garbage so
that we don't hang, fail an assertion, etc.