unity_asset_bundle_extractor: Fix cached blocks using wrong seek
When advancing to the next block via read(), if we pull one that we had
earlier decompressed from cache we did not reset the seek position
within that block back to the start of the block. This could in result
in seemingly corrupt files if they hit certain access patterns.