extract_unity55_shaders: Remove assertion on second tags field
There seem to be two tags fields associated with passes - one is
specifically associated with the pass state, and the other just
generally with the pass. The former is used in many shaders, but the
later seems much rarer, so I had put an assert() on it to catch where it
was used to work out if there are any differences.

This assert triggered on one of the shader assets in the LiSBtS DLC
asset bundles, but the specific pass it was set on contained no
subprograms (no shaders), while other passes within the same shader
asset contained the tags field on the state object and did have
subprograms associated with them.

Weird - I'm still not entirely certain what this signifies, but we can't
fail on it. For now remove the assert and we'll just add two Tags
entries to a pass that contains both (if that is ever possible).