extract_unity55_shaders: Add --verbose option, clean up output
This script is very verbose, which slows down processing fairly
significantly on Windows. Add a --verbose option to show all the output,
and change most of the fields to only show if that option has been
specified.

We still print out some of the messages that show our progress parsing
the data to give an idea that the script is still doing work, and re-jig
some of the output to look nicer for the regular run.