extract_unity55_shaders: Add second level of verbosity
Showing the progress still outputs a fairly large amount of useless
messages, as some of the shader assets contain many hundreds of
individual shaders. Add a second level of verbosity, so we show only a
summary of the number of shaders by default, progress with one level of
verbosity, and a raw dump of all the parsed fields with two levels of
verbosity.