extract_unity55_shaders: Skip decompressing unwanted platform blobs
If the user passes --type=d3d11 or similar, we can skip decompressing
the platform blobs for unrelated platforms to save time. Add a mapping
of identified platform IDs to their corresponding APIs, and reuse
extract_unity53_shaders code to also map that to the conceptual APIs for
filtering.

If a platform does not appear in this map we will still decompress that
platform blob and filtering will be delayed until the individual
shaders, but we can skip any we have identified from the outset.