LiSBtS: Add debugging shaders
Used these to examine the object highlight shader in more detail, and
concluded that it is now working "correctly" (though am yet to decide
whether to change it anyway).

The random noise inputs match in both eyes, as does the jitter offset,
etc. Looking closer, the outline offsets are being applied mostly the
same in both eyes after correcting the view-space coordinate, but this
jitter does depend on the depth in each eye, which is where the slight
differences are coming from.

There is also a texture being applied to the outline, which is dependent
on the depth of the surface behind it. This is the case in 2D as well,
it is just a little more obvious in 3D.

Overall, the effect works OK - could be better, but improving it might
be a lower priority.