Unity 5.5 / LiSBtS: Add d3dx.ini stanza for directional lighting shader
This shader was added for DFC, but it turns out to be used in this game
as well, suggesting that this may now be the Unity 5.5 directional
lighting shader. We've already got the fixed shader in the template, but
we were missing the stanzas to copy UnityPerCameraRare with the
projection matrix out of it, which was causing the physical lighting in
LiSBtS to be messed up.

Add the stanzas to fix it.