Unity 5 / DX11: Reset copy limits from ClearRTV to fix reflections in LiSBtS
The reflections in glass were messed up in this game, and earlier
experiments with the object highlight shader indicated that matrix we
are using for world transformations was not valid, and this is because
resetting the copy & execution limits from the ClearDepthStencilView
call is insufficient in this game.

Resetting these limits from the ClearRenderTargetView call does the
trick, so to be on the safe side change the template to reset them from
both.

This may also explain the broken light glow in the cutscene with Frank -
will need to go back and recheck that.