LiSBtS: Implement shaders for auto-convergence
This uses a brand new feature in 3DMigoto 1.2.69 to automatically adjust
the convergence from shaders on the GPU.

This implements shaders that sample a number of points on the depth
buffer and progressively downscales it to find the closest point to the
camera, and then sets the convergence to match that value.

Two IniParam values are used to set the minimum and maximum auto
convergence that it may choose, to prevent the convergence being set
insanely high when nothing is being drawn (which would be bad, since
that would then mean that everything is drawn off-screen, so there will
be nothing on the depth buffer for the auto-convergence to find to
reduce it).