LiSBtS: Add slow and instant convergence thresholds
It's a little disorienting with the convergence auto-adjusting on every
frame, so add two thresholds between the target and current convergence
that must be exceeded to activate it - a smaller threshold that
activates a slower auto-convergence transition, and a larger threshold
that instantly sets the automatic convergence.

We can probably make further improvements - perhaps we could use a
better method to detect scene transitions to trigger an instant
convergence change, and we probably will need to take the non-linear
nature of depth into account to make this work well for objects at
different distances.