LiSBtS: Improve auto-convergence popout bias formula for different screen sizes + scenes
Applying a fixes convergence bias to give some popout doesn't tend to
work very well - calibrating it for a small amount of pop out in a
regular scene with Chloe ends up with way too much pop out in other
scenes, like the D&D scene when the camera is looking right at the
characters on the board.

A better formula would be one that sets a maximum distance that an
object can be popped out of the screen, regardless of the distance it is
from the camera in game, and finds the convergence that makes that
happen.

This is an experimental formula to do just that. Rather than specifying
a convergence bias, it specifies a bias to the resulting X adjustment,
because this will define how far out of the screen an object appears to
be regardless of its depth. This formula is derived from the nvidia
formula, and in my testing seems to work pretty well on both my 27"
monitor and 120" projector in both regular gameplay and the D&D scene.