LiSBtS: Use proper timing for auto-convergence slow adjustment
Rather than setting the amount of convergence that can be changed as a
per frame value and therefore dependent on framerate, specify it in
terms of convergence/second, and use the new timing feature in 3DMigoto
to make it independent of the framerate. Note that due to the fact that
there is a delay before the convergence is changed, we need to multiply
the rate by 2 to compensate - this may not be accurate in all cases, but
will probably work well enough for now.

This also tweaks the thresholds for triggering the auto-convergence to
use separate thresholds for too close or too far.