LiSBtS / auto-convergence: Implement an anti-judder countermeasure
In some edge cases, something very close to the camera may trigger
auto-convergence to lower the convergence, but in doing so that object
may be moved off-screen. The auto-convergence would then analyse the
next frame and decide that the convergence is now too low and raise it,
bringing the object back on-screen, leading to a judder between two or
more convergence values.

This implements a countermeasure to try to prevent this. The
auto-convergence shader writes the last four auto-convergence values to
a state buffer, and if it tries to set the convergence very close to one
of those values it will abort.

This seems to work pretty well, though the abort may mean that sometimes
the object is left on screen, and we probably want it off.