LiSBtS: Allow user to adjust popout bias with standard convergence keys
To make the auto-convergence easier to adjust, allow the user to use the
standard convergence hotkeys to adjust the convergence.

Since the auto-convergence doesn't operate on convergence, but rather a
popout bias, when it detects the user is attempting to adjust the
convergence, calculate the equivalent popout bias that corresponds to
that and save that to the auto-convergence' state buffer.